Move SwaggerError to bravado.mapping.exception
