+++ bravado/mapping/param.py
@@ -1,7 +1,7 @@
 import urllib
 import simplejson as json
 
-from bravado.exception import SwaggerError
+from bravado.mapping.exception import SwaggerError
 from bravado.http_client import APP_JSON
 from bravado.mapping.marshal import marshal_schema_object
